# perf: inline BeginTransaction with first statement in the transaction