--- a/lib/spanner_client_ext.rb
+++ b/lib/spanner_client_ext.rb
@@ -95,6 +95,10 @@ module Google
         end
       end
 
+      class Results
+        attr_reader :metadata
+      end
+
       class Transaction
         attr_accessor :seqno, :commit
       end