--- a/test/activerecord_spanner_mock_server/inline_begin_transaction_test.rb
+++ b/test/activerecord_spanner_mock_server/inline_begin_transaction_test.rb
@@ -9,22 +9,25 @@
 require_relative "./base_spanner_mock_server_test"
 
 module MockServerTests
-  class AbortedTransactionTest < BaseSpannerMockServerTest
-    def test_read_write_transaction_without_abort_does_not_retry
-      register_insert_singer_result
+  class InlineTransactionTest < BaseSpannerMockServerTest
+    def test_read_write_transaction_uses_inlined_begin
+      insert_sql = register_insert_singer_result
 
       ActiveRecord::Base.transaction do
         Singer.create(first_name: "Dave", last_name: "Allison")
       end
 
-      # There should only be one transaction. The BeginTransaction should be inlined with the first request.
+      # There should be no explicit BeginTransaction request.
       begin_transaction_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::BeginTransactionRequest }
       assert_empty begin_transaction_requests
+      sql_request = @mock.requests.select { |req| req.is_a?(Google::Cloud::Spanner::V1::ExecuteSqlRequest) && req.sql == insert_sql }.first
+      assert sql_request.transaction&.begin&.read_write
+
       commit_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::CommitRequest }
       assert_equal 1, commit_requests.length
     end
 
-    def test_read_write_transaction_aborted_dml_is_automatically_retried
+    def test_read_write_transaction_aborted_dml_is_automatically_retried_with_inline_begin
       insert_sql = register_insert_singer_result
 
       already_aborted = false
@@ -36,18 +39,20 @@ module MockServerTests
       end
 
       # There should be two transaction attempts, two ExecuteSqlRequests and only one commit.
-      # The BeginTransaction should be inlined with the first request.
+      # Both transaction attempts should use an inlined BeginTransaction.
       begin_transaction_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::BeginTransactionRequest }
       assert_empty begin_transaction_requests
       sql_requests = @mock.requests.select { |req| req.is_a?(Google::Cloud::Spanner::V1::ExecuteSqlRequest) && req.sql == insert_sql }
       assert_equal 2, sql_requests.length
+      sql_requests.each { |req| assert req.transaction&.begin&.read_write }
+
       commit_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::CommitRequest }
       assert_equal 1, commit_requests.length
       rollback_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::RollbackRequest }
       assert_empty rollback_requests
     end
 
-    def test_read_write_transaction_aborted_query_is_automatically_retried
+    def test_read_write_transaction_aborted_query_is_automatically_retried_with_inline_begin
       select_sql = register_singer_find_by_id_result
 
       already_aborted = false
@@ -60,13 +65,14 @@ module MockServerTests
       assert_empty begin_transaction_requests
       sql_requests = @mock.requests.select { |req| req.is_a?(Google::Cloud::Spanner::V1::ExecuteSqlRequest) && req.sql == select_sql }
       assert_equal 2, sql_requests.length
+      sql_requests.each { |req| assert req.transaction&.begin&.read_write }
       commit_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::CommitRequest }
       assert_equal 1, commit_requests.length
       rollback_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::RollbackRequest }
       assert_empty rollback_requests
     end
 
-    def test_read_write_transaction_aborted_commit_is_automatically_retried
+    def test_read_write_transaction_aborted_commit_is_automatically_retried_with_inline_begin
       select_sql = register_singer_find_by_id_result
 
       already_aborted = false
@@ -79,17 +85,19 @@ module MockServerTests
       assert_empty begin_transaction_requests
       sql_requests = @mock.requests.select { |req| req.is_a?(Google::Cloud::Spanner::V1::ExecuteSqlRequest) && req.sql == select_sql }
       assert_equal 2, sql_requests.length
+      sql_requests.each { |req| assert req.transaction&.begin&.read_write }
       commit_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::CommitRequest }
       assert_equal 2, commit_requests.length
       rollback_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::RollbackRequest }
       assert_empty rollback_requests
     end
 
-    def test_implicit_transaction_aborted_single_insert_is_automatically_retried
+    def test_implicit_transaction_aborted_single_insert_is_automatically_retried_without_inline_begin
       # This will abort the next transaction that is started on the mock server.
       @mock.abort_next_transaction
       # The following statement will automatically start a transaction, although it is not in a transaction block.
       # The transaction is automatically retried if the transaction is aborted.
+      # It will not inline the BeginTransaction option, as it only uses mutations.
       Singer.create(first_name: "Dave", last_name: "Allison")
 
       # There should be two transaction attempts.
@@ -101,12 +109,13 @@ module MockServerTests
       assert_empty rollback_requests
     end
 
-    def test_implicit_transaction_aborted_batch_insert_is_automatically_retried
+    def test_implicit_transaction_aborted_batch_insert_is_automatically_retried_without_inline_begin
       @mock.abort_next_transaction
       Singer.create([{first_name: "Dave", last_name: "Allison"}, {first_name: "Alice", last_name: "Becker"}])
 
       # The batch will be inserted as one transaction containing two mutations. The first attempt will abort, and
       # the second will succeed.
+      # It will not inline the BeginTransaction option, as it only uses mutations.
       begin_transaction_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::BeginTransactionRequest }
       assert_equal 2, begin_transaction_requests.length
       commit_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::CommitRequest }