--- a/test/activerecord_spanner_mock_server/inline_begin_transaction_test.rb
+++ b/test/activerecord_spanner_mock_server/inline_begin_transaction_test.rb
@@ -124,6 +124,152 @@ module MockServerTests
       assert_empty rollback_requests
     end
 
+    def test_unhandled_error_on_first_statement_retries_transaction_and_fails
+      insert_sql = register_insert_singer_result
+      # Push the same error twice, so the same error will be returned both during the initial
+      # attempt and during the retry.
+      @mock.push_error \
+        insert_sql, \
+        GRPC::BadStatus.new(
+          GRPC::Core::StatusCodes::FAILED_PRECONDITION,
+          "A row with the given identifier already exists")
+      @mock.push_error \
+        insert_sql, \
+        GRPC::BadStatus.new(
+        GRPC::Core::StatusCodes::FAILED_PRECONDITION,
+        "A row with the given identifier already exists")
+
+      # The entire transaction should be retried if the first statement in a transaction
+      # causes an error with a different error code than Aborted, and the retried transaction
+      # should use an explicit BeginTransaction RPC. This is necessary in order to include
+      # the statement that caused the error in the transaction, as the error could be an
+      # indication of the state of the database during the transaction. A unique key constraint
+      # violation for example indicates that a record with a specific key value exists.
+      err = assert_raises ActiveRecord::StatementInvalid do
+        ActiveRecord::Base.transaction do
+          Singer.create(first_name: "Dave", last_name: "Allison")
+        end
+      end
+      assert err.cause.is_a?(Google::Cloud::FailedPreconditionError)
+
+      # There should be one explicit BeginTransaction request.
+      begin_transaction_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::BeginTransactionRequest }
+      assert_equal 1, begin_transaction_requests.length
+      sql_requests = @mock.requests.select { |req| req.is_a?(Google::Cloud::Spanner::V1::ExecuteSqlRequest) && req.sql == insert_sql }
+      assert_equal 2, sql_requests.length
+      assert sql_requests[0].transaction&.begin&.read_write
+      assert sql_requests[1].transaction&.id
+
+      commit_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::CommitRequest }
+      assert_empty commit_requests
+    end
+
+    def test_handled_error_on_first_statement_retries_transaction_and_succeeds
+      insert_sql = register_insert_singer_result
+      # Push the same error twice, so the same error will be returned both during the initial
+      # attempt and during the retry.
+      @mock.push_error \
+        insert_sql, \
+        GRPC::BadStatus.new(
+        GRPC::Core::StatusCodes::FAILED_PRECONDITION,
+        "A row with the given identifier already exists")
+      @mock.push_error \
+        insert_sql, \
+        GRPC::BadStatus.new(
+        GRPC::Core::StatusCodes::FAILED_PRECONDITION,
+        "A row with the given identifier already exists")
+
+      attempts = 0
+      ActiveRecord::Base.transaction do
+        begin
+          Singer.create(first_name: "Dave", last_name: "Allison")
+        rescue ActiveRecord::StatementInvalid
+          raise if (attempts += 1) > 1
+          # Ignore the error and retry the statement.
+          retry
+        end
+      end
+
+      # There should be one explicit BeginTransaction request.
+      begin_transaction_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::BeginTransactionRequest }
+      assert_equal 1, begin_transaction_requests.length
+      sql_requests = @mock.requests.select { |req| req.is_a?(Google::Cloud::Spanner::V1::ExecuteSqlRequest) && req.sql == insert_sql }
+      # There should be 3 ExecuteSqlRequests:
+      # 1. The initial attempt that includes a BeginTransaction. This statement will fail and be
+      #    retried after an explicit BeginTransaction request.
+      # 2. The retried attempt after the BeginTransaction RPC. This statement will also fail, as
+      #    the error has been added twice.
+      # 3. The error handler in the transaction block will retry the statement once more. This time
+      #    the statement will succeed, as the error has only been pushed twice.
+      assert_equal 3, sql_requests.length
+      assert sql_requests[0].transaction&.begin&.read_write
+      assert sql_requests[1].transaction&.id
+      assert_equal sql_requests[1].transaction&.id, sql_requests[2].transaction&.id
+
+      commit_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::CommitRequest }
+      assert_equal 1, commit_requests.length
+      assert_equal sql_requests[1].transaction&.id, commit_requests[0].transaction_id
+    end
+
+    def test_transient_error_on_first_statement_retries_transaction_and_succeeds
+      insert_sql = register_insert_singer_result
+      # Push the same error only once, so the retried statement will succeed.
+      @mock.push_error \
+        insert_sql, \
+        GRPC::BadStatus.new(
+        GRPC::Core::StatusCodes::RESOURCE_EXHAUSTED,
+        "Too many requests")
+
+      ActiveRecord::Base.transaction do
+        # This statement will fail once, and then be retried internally after an explicit
+        # BeginTransaction RPC. The second attempt will succeed as the error was only pushed once.
+        Singer.create(first_name: "Dave", last_name: "Allison")
+      end
+
+      # There should be one explicit BeginTransaction request.
+      begin_transaction_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::BeginTransactionRequest }
+      assert_equal 1, begin_transaction_requests.length
+      sql_requests = @mock.requests.select { |req| req.is_a?(Google::Cloud::Spanner::V1::ExecuteSqlRequest) && req.sql == insert_sql }
+      assert_equal 2, sql_requests.length
+      assert sql_requests[0].transaction&.begin&.read_write
+      assert sql_requests[1].transaction&.id
+
+      commit_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::CommitRequest }
+      assert_equal 1, commit_requests.length
+      assert_equal sql_requests[1].transaction&.id, commit_requests[0].transaction_id
+    end
+
+    def test_error_on_first_statement_and_on_begin_transaction
+      insert_sql = register_insert_singer_result
+      @mock.push_error \
+        insert_sql, \
+        GRPC::BadStatus.new(GRPC::Core::StatusCodes::FAILED_PRECONDITION,
+                            "A row with the given identifier already exists")
+      @mock.push_error \
+        "begin_transaction", \
+        GRPC::BadStatus.new(GRPC::Core::StatusCodes::PERMISSION_DENIED, "Not permitted")
+
+      err = assert_raises ActiveRecord::StatementInvalid do
+        ActiveRecord::Base.transaction do
+          # This statement will fail and cause an internal retry with an explicit BeginTransaction
+          # RPC. That RPC invocation will fail, but the original error from the SQL statement
+          # will be the error that is returned.
+          Singer.create(first_name: "Dave", last_name: "Allison")
+        end
+      end
+      assert err.cause.is_a?(Google::Cloud::FailedPreconditionError)
+
+      # There should be one explicit BeginTransaction request.
+      begin_transaction_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::BeginTransactionRequest }
+      assert_equal 1, begin_transaction_requests.length
+      sql_requests = @mock.requests.select { |req| req.is_a?(Google::Cloud::Spanner::V1::ExecuteSqlRequest) && req.sql == insert_sql }
+      assert_equal 1, sql_requests.length
+      assert sql_requests[0].transaction&.begin&.read_write
+
+      commit_requests = @mock.requests.select { |req| req.is_a? Google::Cloud::Spanner::V1::CommitRequest }
+      assert_empty commit_requests
+    end
+
     def register_insert_singer_result
       sql = "INSERT INTO `singers` (`first_name`, `last_name`, `id`) VALUES (@p1, @p2, @p3)"
       @mock.put_statement_result sql, StatementResult.new(1)