Replaces the `BeginTransaction` RPC invocation at the beginning of a read/write transaction with an inlined `BeginTransaction` option on the first statement in the transaction. This reduces the number of roundtrips by one for read/write transactions that execute at least one SQL statement (DML or query).

Transactions that only execute one or more mutations still need to execute a `BeginTransaction` RPC before the `Commit`.

The [inline_begin_transaction_test.rb](https://github.com/googleapis/ruby-spanner-activerecord/blob/35df01ca87d72cdbbf3d1a3e63736c88462c48cc/test/activerecord_spanner_mock_server/inline_begin_transaction_test.rb) file is a good starting point for reviewing this PR, as it gives an overview of how the provider behaves for different scenarios.

## Benchmarks

```
Benchmarks for ActiveRecord
                                                                                  user     system      total        real
Select one row:                                                               0.005923   0.000108   0.006031 (  0.020545)
Save one row with fetch after:                                                0.004887   0.000050   0.004937 (  0.025885)
Select and update 1 row in transaction using mutation:                        0.005404   0.000000   0.005404 (  0.075397)
Select and update 1 row in transaction using DML:                             0.005592   0.000174   0.005766 (  0.043225)
Select and update 25 rows in transaction using mutation:                      0.034073   0.008113   0.042186 (  0.151752)
Select and update 25 rows in transaction using DML:                           0.057145   0.008064   0.065209 (  0.298057)
Create 100 albums using mutations:                                            0.049477   0.004085   0.053562 (  0.098786)
Create 100 albums using DML:                                                  0.176151   0.008699   0.184850 (  0.688414)
Select and iterate over 100 singers:                                          0.007212   0.000000   0.007212 (  0.028932)
Select and iterate over 100 singers in a read-only transaction:               0.008563   0.000222   0.008785 (  0.039447)
Select and iterate over 100 singers in a read/write transaction:              0.005572   0.003742   0.009314 (  0.030663)
                                                                                  user     system      total        real
Total execution time (1):                                                     0.058740   0.000000   0.058740 (  0.261739)
Total execution time (5):                                                     0.204751   0.007962   0.212713 (  0.345838)
Total execution time (10):                                                    0.447394   0.015949   0.463343 (  0.561728)
Total execution time (25):                                                    0.983090   0.095282   1.078372 (  1.255142)
Total execution time (50):                                                    2.211445   0.116437   2.327882 (  2.576518)
Total execution time (100):                                                   4.391837   0.316523   4.708360 (  5.131198)
Total execution time (200):                                                   9.795164   0.647298  10.442462 ( 11.102745)
Total execution time (400):                                                  17.320740   1.470954  18.791694 ( 30.548032)
Benchmarks for Spanner client
                                                                                  user     system      total        real
Select one row:                                                               0.003363   0.000000   0.003363 (  0.031631)
Save one row with fetch after:                                                0.002478   0.000139   0.002617 (  0.016598)
Select and update 1 row in transaction using mutation:                        0.002654   0.003920   0.006574 (  0.054519)
Select and update 1 row in transaction using DML:                             0.006363   0.000000   0.006363 (  0.046582)
Select and update 25 rows in transaction using mutation:                      0.029641   0.003963   0.033604 (  0.180197)
Select and update 25 rows in transaction using DML:                           0.054942   0.008510   0.063452 (  0.367085)
Create 100 albums using mutations:                                            0.010886   0.000207   0.011093 (  0.039520)
Create 100 albums using DML:                                                  0.031594   0.003975   0.035569 (  0.348762)
Select and iterate over 100 singers:                                          0.004819   0.000058   0.004877 (  0.027926)
Select and iterate over 100 singers in a read-only transaction:               0.005611   0.000000   0.005611 (  0.043016)
Select and iterate over 100 singers in a read/write transaction:              0.005407   0.000000   0.005407 (  0.041318)
                                                                                  user     system      total        real
Total execution time (1):                                                     0.061344   0.003226   0.064570 (  0.270657)
Total execution time (5):                                                     0.127943   0.003408   0.131351 (  0.295201)
Total execution time (10):                                                    0.224252   0.039778   0.264030 (  0.626101)
Total execution time (25):                                                    0.502620   0.080572   0.583192 (  0.689937)
Total execution time (50):                                                    1.009005   0.161462   1.170467 (  1.351502)
Total execution time (100):                                                   2.238698   0.343673   2.582371 (  2.858914)
Total execution time (200):                                                   5.099566   0.746336   5.845902 (  8.837476)
Total execution time (400):                                                   9.245205   1.741330  10.986535 ( 59.127166)
```

### Side-by-side
The table below contains the same results as above, but presented side-by-side for a direct comparison between the normal Spanner client and the ActiveRecord provider. The figures shown are the real execution times, i.e. including network time. The `ActiveRecord (original)` times are taken from the original benchmark run (before this change). The `ActiveRecord (new)` column contains the measurements after this change.

| Benchmark | ActiveRecord (original) | Spanner client | ActiveRecord (new) |
|------------|------------------------|----------------|---------------------|
| Select one row | 0.011054 | 0.031631 | 0.020545 |
| Save one row with fetch after | 0.030973 | 0.016598 | 0.025885 |
| Select and update 1 row in transaction using mutation | 0.178689 | 0.054519 | 0.075397 |
| Select and update 1 row in transaction using DML | 0.139983 | 0.046582 | 0.043225 |
| Select and update 25 rows in transaction using mutation | 0.304142 | 0.151752 | 0.151752 |
| Select and update 25 rows in transaction using DML | 0.649054 | 0.367085 | 0.298057 |
| Create 100 albums using mutations | 0.220688 | 0.039520 | 0.098786 |
| Create 100 albums using DML | 1.022725 | 0.348762 | 0.688414 |
| Select and iterate over 100 singers | 0.115998 | 0.027926 | 0.028932 |
| Select and iterate over 100 singers in a read-only transaction | 0.050298 | 0.043016 | 0.039447 |
| Select and iterate over 100 singers in a read/write transaction | 0.054801 | 0.041318 | 0.030663 |
| Total execution time (1) | 0.643255 | 0.270657 | 0.261739 |
| Total execution time (5) | 0.754162 | 0.295201 | 0.345838 |
| Total execution time (10) | 0.909938 | 0.626101 | 0.561728 |
| Total execution time (25) | 1.925742 | 0.689937 | 1.255142 |
| Total execution time (50) | 3.396065 | 1.351502 | 2.576518 |
| Total execution time (100) | 6.944054 | 2.858914 | 5.131198 |
| Total execution time (200) | 14.459822 | 8.837476 | 11.102745 |
| Total execution time (400) | 30.333424 | 59.127166 | 30.548032 |