@hengfengli 2021-11-19
Thanks for making this. It looks like that it has some improvements. Do you know why sometimes, the new code is even better than the spanner client, e.g., `Select and iterate over 100 singers in a read/write transaction`?