@olavloite 2021-11-19
> Thanks for making this. It looks like that it has some obvious improvements. Do you know why sometimes, the new code is even better than the spanner client, e.g., `Select and iterate over 100 singers in a read/write transaction`?

It can sometimes be down to pure coincidence/luck, as the benchmark times include network roundtrip times and possibly also network time and other slightly random processing time on the backend. But all benchmark scenarios that use a read/write transaction and at least one SQL statement __can__ have an advantage over the normal client, as the normal client needs to get a write-prepared session from the pool, while the ActiveRecord provider will always inline the BeginTransaction option with the SQL statement. I'm not really familiar with the session pool in the Ruby client, but from what I can see in the implementation I think the difference is down to the following:
1. The benchmarks use a default session pool for the normal Spanner client. The default session pool has `MinSessions = 10` and `WriteFraction = 0.3`. That means that the initial session pool will have 3 sessions that are prepared for read/write.
2. As the benchmarks execute more than 3 read/write transactions, there is a possibility that there is no read/write prepared session available at the moment that the `Select and iterate over 100 singers in a read/write transaction` benchmark is being executed. This means that the benchmark might have to wait for the pool to execute a `BeginTransaction` RPC on a session before it gets a session from the pool.
3. In addition, the session pool uses a background executor to prepare sessions with read/write transactions. The number of threads in this executor depends on the number of processors in the VM/server that executes the application (but it is never less than 2). This can cause queueing, especially in a scenario with a very high level of concurrency, as the number of transactions that can be created is limited by the number of worker threads. I think that this is what you are seeing in the `Total execution time (400)` scenario.